refactor(common): move lock providers to org.apache.hudi.core.transaction.lock
Relocates InProcessLockProvider and NoopLockProvider out of
org.apache.hudi.client.transaction.lock in hudi-common, pre-staging the
future hudi-common/hudi-core module split so the eventual module move
stays FQN-stable.

- InProcessLockProvider: the old FQN is a documented, user-configured
  value of hoodie.write.lock.provider resolved via reflection, so a
  deprecated stub is kept at the old FQN, marked with the new
  @CompatAlias annotation (hudi-io) recording the replacement and the
  1.3.0 deprecation release. The stub preserves the reflective
  (LockConfiguration, StorageConfiguration) constructor contract.
- NoopLockProvider: internal-only (upgrade/downgrade nested-lock path),
  moved without a stub.
- Sites that string-compare the configured lock provider class now use
  InProcessLockProvider.isInProcessLockProvider(), which accepts both
  the current and legacy FQN: the 1ms clock-skew inference in
  HoodieTimeGeneratorConfig, the metadata-table multi-writer guard in
  HoodieMetadataWriteUtils, and the single-writer auto-adjust in
  HoodieWriteConfig. Without this, configs carrying the legacy FQN
  would silently lose those behaviors.

Behavior-preserving otherwise. Legacy-FQN test configs are deliberately
left in place as regression coverage for the stub path; a reflective
load test and a legacy-FQN clock-skew row are added.
