+++ hudi-client/hudi-client-common/src/test/java/org/apache/hudi/table/TestHoodieTable.java
@@ -20,7 +20,6 @@ package org.apache.hudi.table;
 
 import org.apache.hudi.avro.model.HoodieRollbackPlan;
 import org.apache.hudi.client.transaction.TransactionManager;
-import org.apache.hudi.client.transaction.lock.InProcessLockProvider;
 import org.apache.hudi.common.HoodiePendingRollbackInfo;
 import org.apache.hudi.common.config.HoodieMetadataConfig;
 import org.apache.hudi.common.engine.HoodieEngineContext;
@@ -33,6 +32,7 @@ import org.apache.hudi.common.util.Option;
 import org.apache.hudi.common.util.SerializationUtils;
 import org.apache.hudi.config.HoodieLockConfig;
 import org.apache.hudi.config.HoodieWriteConfig;
+import org.apache.hudi.core.transaction.lock.InProcessLockProvider;
 import org.apache.hudi.index.HoodieIndex;
 import org.apache.hudi.metadata.HoodieTableMetadata;
 import org.apache.hudi.table.storage.HoodieStorageLayout;
